+++ packages/core-js/modules/esnext.math.signbit.js
@@ -1,5 +1,5 @@
 // `Math.signbit` method
-// http://jfbastien.github.io/papers/Math.signbit.html
+// https://github.com/tc39/proposal-Math.signbit
 require('../internals/export')({ target: 'Math', stat: true }, { signbit: function signbit(x) {
   // eslint-disable-next-line no-self-compare
   return (x = +x) != x ? x : x == 0 ? 1 / x == Infinity : x > 0;