+++ packages/core-js/modules/esnext.reflect.get-own-metadata-keys.js
@@ -4,7 +4,7 @@ var ordinaryOwnMetadataKeys = ReflectMetadataModule.keys;
 var toMetaKey = ReflectMetadataModule.key;
 
 // `Reflect.getOwnMetadataKeys` method
-// https://rbuckton.github.io/reflect-metadata/
+// https://github.com/rbuckton/reflect-metadata
 ReflectMetadataModule.exp({ getOwnMetadataKeys: function getOwnMetadataKeys(target /* , targetKey */) {
   return ordinaryOwnMetadataKeys(anObject(target), arguments.length < 2 ? undefined : toMetaKey(arguments[1]));
 } });