+++ packages/core-js/modules/esnext.reflect.get-own-metadata.js
@@ -4,7 +4,7 @@ var ordinaryGetOwnMetadata = ReflectMetadataModule.get;
 var toMetaKey = ReflectMetadataModule.key;
 
 // `Reflect.getOwnMetadata` method
-// https://rbuckton.github.io/reflect-metadata/
+// https://github.com/rbuckton/reflect-metadata
 ReflectMetadataModule.exp({ getOwnMetadata: function getOwnMetadata(metadataKey, target /* , targetKey */) {
   return ordinaryGetOwnMetadata(metadataKey, anObject(target)
     , arguments.length < 3 ? undefined : toMetaKey(arguments[2]));