+++ packages/core-js/modules/esnext.reflect.has-own-metadata.js
@@ -4,7 +4,7 @@ var ordinaryHasOwnMetadata = ReflectMetadataModule.has;
 var toMetaKey = ReflectMetadataModule.key;
 
 // `Reflect.hasOwnMetadata` method
-// https://rbuckton.github.io/reflect-metadata/
+// https://github.com/rbuckton/reflect-metadata
 ReflectMetadataModule.exp({ hasOwnMetadata: function hasOwnMetadata(metadataKey, target /* , targetKey */) {
   return ordinaryHasOwnMetadata(metadataKey, anObject(target)
     , arguments.length < 3 ? undefined : toMetaKey(arguments[2]));