+++ packages/core-js/modules/esnext.symbol.description.js
@@ -1,5 +1,5 @@
 // `Symbol.prototype.description` getter
-// https://tc39.github.io/proposal-Symbol-description/
+// https://github.com/tc39/proposal-Symbol-description
 'use strict';
 var DESCRIPTORS = require('../internals/descriptors');
 var has = require('../internals/has');