+++ packages/core-js/proposals/array-flat-and-flat-map.js
@@ -1,0 +1,2 @@
+require('../modules/esnext.array.flat');
+require('../modules/esnext.array.flat-map');