+++ packages/core-js/proposals/collection-methods.js
@@ -1,0 +1,15 @@
+require('../modules/esnext.map.group-by');
+require('../modules/esnext.map.key-by');
+require('../modules/esnext.map.filter');
+require('../modules/esnext.map.map-keys');
+require('../modules/esnext.map.map-values');
+require('../modules/esnext.map.merge');
+require('../modules/esnext.set.add-all');
+require('../modules/esnext.set.delete-all');
+require('../modules/esnext.set.every');
+require('../modules/esnext.set.filter');
+require('../modules/esnext.set.find');
+require('../modules/esnext.set.join');
+require('../modules/esnext.set.map');
+require('../modules/esnext.set.reduce');
+require('../modules/esnext.set.some');