+++ packages/core-js/proposals/object-from-entries.js
@@ -1,0 +1,2 @@
+require('../modules/es.array.iterator');
+require('../modules/esnext.object.from-entries');