+++ packages/core-js/proposals/pattern-matching.js
@@ -1,0 +1,2 @@
+require('../modules/esnext.symbol.pattern-match');
+require('../modules/esnext.symbol.pattern-value');