+++ packages/core-js/proposals/string-trim-start-end.js
@@ -1,0 +1,2 @@
+require('../modules/esnext.string.trim-left');
+require('../modules/esnext.string.trim-right');