+++ packages/core-js/stage/0.js
@@ -1,10 +1,6 @@
-require('../modules/esnext.asap');
-require('../modules/esnext.math.iaddh');
-require('../modules/esnext.math.isubh');
-require('../modules/esnext.math.imulh');
-require('../modules/esnext.math.umulh');
-require('../modules/esnext.string.at');
-require('../modules/esnext.symbol.pattern-match');
-require('../modules/esnext.symbol.pattern-value');
+require('../proposals/asap');
+require('../proposals/efficient-64-bit-arithmetic');
+require('../proposals/string-at');
+require('../proposals/pattern-matching');
 
 module.exports = require('./1');