+++ packages/core-js/stage/2.js
@@ -1,7 +1,4 @@
-require('../modules/esnext.object.from-entries');
-require('../modules/esnext.set.difference');
-require('../modules/esnext.set.intersection');
-require('../modules/esnext.set.union');
-require('../modules/esnext.set.symmetric-difference');
+require('../proposals/object-from-entries');
+require('../proposals/set-methods');
 
 module.exports = require('./3');