+++ packages/core-js/stage/3.js
@@ -1,9 +1,7 @@
-require('../modules/esnext.array.flat');
-require('../modules/esnext.array.flat-map');
-require('../modules/esnext.global');
-require('../modules/esnext.string.match-all');
-require('../modules/esnext.string.trim-left');
-require('../modules/esnext.string.trim-right');
-require('../modules/esnext.symbol.description');
+require('../proposals/array-flat-and-flat-map');
+require('../proposals/global');
+require('../proposals/string-match-all');
+require('../proposals/string-trim-start-end');
+require('../proposals/symbol-description');
 
 module.exports = require('./4');