+++ CHANGELOG.md
@@ -102,6 +102,7 @@
   - Move all features from ES5, ES2015, ES2016, ES2017 and ES2018 to one namespace for stable ES - it's available as `core-js/es`, all those features in `modules` folder has `es.` prefix.
   - Change prefix for ES proposals from `es7.` to `esnext.`, they no longer available in `core-js/es7`, use `core-js/stage/*` instead of that.
   - Rename `core-js(/library)/fn` to `core-js(-pure)/features` for improve readability.
+  - Add `/proposals/` entry points for allow include all features from one proposal (for example, `core-js/proposals/reflect-metadata`).
   - Allow more granular inclusion of features from `/es/` path (for example, `core-js/es/array/from`).
   - Split typed arrays polyfills. Now you can, for example, load only required method (for example, `core-js/es/typed-array/from`).
   - Extract well-known symbols definition from `es.symbol` module for loading only required features, for example, in MS Edge.